feat(ui): remove old non-react code (#176)
