--- src/traceViewer/web/components/listView.css
@@ -1,44 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-list-view {
-  display: flex;
-}
-
-list-view:focus {
-  outline: none;
-}
-
-list-view.vertical {
-  flex-direction: column;
-  flex: auto;
-}
-
-list-view.horizontal {
-  flex-direction: row;
-}
-
-list-item {
-  display: flex;
-  flex: none;
-  outline: none;
-}
-
-list-view list-view-drop {
-  margin-top: 5px;
-  flex: 0 0 100px;
-  background-color: var(--active-background);
-}
