--- src/traceViewer/web/components/pwCombo.css
@@ -1,110 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-:host(pw-combo) {
-  flex: auto;
-  display: flex;
-  align-items: center;
-  background: var(--active-background);
-  border: 1px solid transparent;
-  padding-right: 5px;
-  height: 22px;
-}
-
-:host(pw-combo:focus-within) {
-  border-color: var(--focus-ring);
-  outline: none;
-}
-
-input {
-  align-self: center;
-  font-size: 13px;
-  width: 100%;
-  -webkit-appearance: none;
-  border: none;
-  color: inherit;
-  background: transparent;
-  padding: 5px;
-}
-
-input:disabled {
-  opacity: 0.6;
-}
-
-input:focus {
-  outline: none;
-}
-
-codicon:hover {
-  color: white;
-  cursor: pointer;
-}
-
-.pw-combo-menu {
-  border: 1px solid var(--separator);
-  background-color: var(--active-background);
-  position: absolute;
-  top: var(--top);
-  max-height: calc(100% - var(--top));
-  overflow: auto;
-  border-radius: 3px;
-  font-size: 13px;
-}
-
-.pw-combo-menu > list-view {
-  overflow-x: hidden;
-  align-items: stretch;
-  max-width: 100%;
-}
-
-.pw-combo-menu .pw-combo-menu-item {
-  padding: 5px;
-  cursor: pointer;
-  align-items: stretch;
-  max-width: 100%;
-}
-
-.pw-combo-menu .pw-combo-menu-item > span {
-  flex: auto;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.pw-combo-menu .pw-combo-menu-item > span.subtitle {
-  flex: none;
-  max-width: 100%;
-  padding-left: 10px;
-  color: var(--gray);
-  text-align: end;
-}
-
-.pw-combo-menu .pw-combo-menu-item:hover {
-  background-color: var(--selection);
-}
-
-.vbox {
-  display: flex;
-  flex-direction: column;
-  flex: auto;
-  position: relative;
-}
-
-.hbox {
-  display: flex;
-  flex: auto;
-  position: relative;
-}
