--- src/traceViewer/web/components/pwExpandable.css
@@ -1,41 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-:host {
-  display: flex;
-  flex-direction: column;
-}
-
-.title {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  white-space: nowrap;
-}
-
-pw-small-button {
-  margin-right: 4px;
-}
-
-:host .body {
-  display: none;
-  flex: auto;
-  margin: 5px 0 5px 20px;
-}
-
-:host([expanded]) .body {
-  display: flex;
-}
