--- src/traceViewer/web/components/splitView.css
@@ -1,113 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-split-view {
-  flex: auto;
-  display: flex;
-  overflow: hidden;
-  --sidebar-size: 300px;
-  position: relative;
-}
-
-split-view.horizontal {
-  flex-direction: row;
-}
-
-split-view.vertical {
-  flex-direction: column;
-}
-
-split-view split-view-sidebar {
-  overflow: hidden;
-  flex: 0 0 var(--sidebar-size);
-  display: flex;
-  position: relative;
-}
-
-split-view.sidebar-hidden > split-view-sidebar {
-  display: none;
-}
-
-split-view.sidebar-hidden > resizer-nob {
-  display: none;
-}
-
-split-view split-view-main {
-  overflow: hidden;
-  flex: auto;
-  display: flex;
-  position: relative;
-}
-
-split-view .is-being-resized {
-  pointer-events: none;
-}
-
-resizer-nob {
-  position: absolute;
-  font-family: monospace;
-  color: white;
-  display: flex;
-  flex: none;
-  align-items: center;
-  justify-content: center;
-  z-index: 100;
-}
-
-resizer-nob.top {
-  top: calc(var(--sidebar-size) - var(--layout-gap));
-}
-
-resizer-nob.right {
-  right: calc(var(--sidebar-size) - var(--layout-gap));
-}
-
-resizer-nob.bottom {
-  bottom: calc(var(--sidebar-size) - var(--layout-gap));
-}
-
-resizer-nob.left {
-  left: calc(var(--sidebar-size) - var(--layout-gap));
-}
-
-resizer-nob.horizontal {
-  cursor: row-resize;
-  width: 11px;
-  padding: 5px 0;
-  width: 100%;
-}
-
-resizer-nob.vertical {
-  cursor: col-resize;
-  width: 11px;
-  padding: 0 5px;
-  height: 100%;
-}
-
-resizer-nob::before {
-  color: transparent;
-  content: 'a';
-}
-
-resizer-nob.vertical::before {
-  width: 1px;
-  height: 100%;
-}
-
-resizer-nob.horizontal::before {
-  height: 1px;
-  width: 100%;
-}
