--- src/traceViewer/web/components/tabbedPane.css
@@ -1,78 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-tabbed-pane {
-  display: flex;
-  flex: auto;
-  overflow: hidden;
-}
-
-tabbed-pane tab-content {
-  display: flex;
-  flex: auto;
-  overflow: hidden;
-}
-
-tabbed-pane tab-strip {
-  flex: auto;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-tabbed-pane tab-strip > list-view {
-  height: 34px;
-}
-
-tabbed-pane tab-element {
-  padding: 2px 6px 0 6px;
-  margin-right: 4px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  user-select: none;
-  border-bottom: 3px solid transparent;
-  width: 80px;
-}
-
-tabbed-pane tab-label {
-  max-width: 250px;
-  white-space: pre;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: inline-block;
-}
-
-tabbed-pane list-item.selected tab-element {
-  border-bottom-color: var(--color);
-}
-
-tabbed-pane tab-close {
-  visibility: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 -4px 0 2px;
-}
-
-tabbed-pane tab-element:hover {
-  font-weight: 600;
-}
-
-tabbed-pane tab-element:hover tab-close {
-  visibility: visible;
-}
