--- src/traceViewer/web/components/toolbarView.css
@@ -1,61 +1,0 @@
-/*
-  Copyright (c) Microsoft Corporation.
- 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-toolbar-view {
-  height: 34px;
-  display: flex;
-  white-space: nowrap;
-  padding: var(--layout-gap) var(--layout-gap);
-}
-
-toolbar-view toolbar-item {
-  height: 24px;
-  display: flex;
-  align-items: center;
-  flex: none;
-  text-transform: uppercase;
-  cursor: pointer;
-  user-select: none;
-  padding: 0 5px;
-  margin: 0 2px;
-}
-
-toolbar-view toolbar-item.toggled {
-  background-color: var(--active-background);
-  color: white;
-}
-
-toolbar-view toolbar-item.disabled {
-  opacity: 0.5;
-  cursor: default;
-}
-
-toolbar-view toolbar-item:not(.disabled):hover {
-  color: #ffffff;
-}
-
-toolbar-view toolbar-item:not(.disabled):hover .codicon {
-  color: #ffffff;
-}
-
-toolbar-view toolbar-icon {
-  display: flex;
-  transition: transform 50ms ease;
-}
-
-toolbar-view toolbar-item:not(.disabled):active toolbar-icon {
-  transform: scale(1.272019649)
-}
